examples/train_federated_model.py
"""
Example: Train Federated SRL Model
"""

import sys
import os
sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

import torch
import numpy as np
import yaml
from src.data.synthetic_generator import SyntheticDataGenerator
from src.training.federated_trainer import FederatedTrainer
from src.evaluation.evaluator import Evaluator


def main():
    """Main training function"""
    
    print("="*80)
    print("Adaptive Multi-Modal AI Framework for Self-Regulated Learning")
    print("Federated Deep Reinforcement Learning Training")
    print("="*80)
    
    # Load configuration
    config_path = 'configs/default.yaml'
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    
    # Set random seed for reproducibility
    random_seed = config.get('random_seed', 42)
    torch.manual_seed(random_seed)
    np.random.seed(random_seed)
    
    # Generate synthetic data
    print("\n[1/5] Generating synthetic data...")
    generator = SyntheticDataGenerator(
        num_institutions=config['training']['federated']['num_institutions'],
        num_learners=config['data']['num_learners'],
        num_days=config['data']['num_days'],
        domains=config['data']['domains'],
        random_seed=random_seed
    )
    
    dataset = generator.generate()
    
    # Save dataset
    os.makedirs('data', exist_ok=True)
    generator.save(dataset, 'data/synthetic_dataset.pkl')
    
    # Prepare training data
    print("\n[2/5] Preparing training data...")
    training_data = generator.prepare_for_training(dataset)
    
    # Calculate institution weights based on data size
    institution_weights = [len(data['states']) for data in training_data]
    total_samples = sum(institution_weights)
    institution_weights = [w / total_samples for w in institution_weights]
    
    print(f"Institution weights: {[f'{w:.3f}' for w in institution_weights]}")
    
    # Initialize federated trainer
    print("\n[3/5] Initializing federated trainer...")
    trainer = FederatedTrainer(
        num_institutions=config['training']['federated']['num_institutions'],
        state_dim=config['training']['rl']['state_dim'],
        action_dim=config['training']['rl']['action_dim'],
        num_rounds=config['training']['federated']['num_rounds'],
        local_epochs=config['training']['federated']['local_epochs'],
        epsilon=config['training']['privacy']['epsilon'],
        delta=config['training']['privacy']['delta'],
        noise_multiplier=config['training']['privacy']['noise_multiplier'],
        clip_norm=config['training']['privacy']['clip_norm']
    )
    
    # Train model
    print("\n[4/5] Training federated model...")
    history = trainer.train(
        institutional_data=training_data,
        institution_weights=institution_weights
    )
    
    # Save trained model
    os.makedirs('models', exist_ok=True)
    model_path = 'models/federated_srl_model.pt'
    trainer.save(model_path)
    print(f"\nModel saved to {model_path}")
    
    # Evaluate model
    print("\n[5/5] Evaluating model...")
    evaluator = Evaluator(output_dir='results')
    
    # Evaluate on test data (using first institution as example)
    test_metrics = trainer.evaluate(training_data[0], institution_id=0)
    
    print("\nTest Metrics:")
    for key, value in test_metrics.items():
        print(f"  {key}: {value:.4f}")
    
    # Plot convergence
    print("\nGenerating convergence plots...")
    evaluator.plot_convergence(
        history,
        save_path='results/convergence.png'
    )
    
    # Generate comprehensive report
    all_metrics = {
        'training': history,
        'test': test_metrics,
        'privacy': {
            'epsilon': config['training']['privacy']['epsilon'],
            'delta': config['training']['privacy']['delta'],
            'privacy_score': 0.95  # Computed based on epsilon
        }
    }
    
    evaluator.generate_report(all_metrics, save_dir='results')
    
    print("\n" + "="*80)
    print("Training completed successfully!")
    print("Results saved to 'results/' directory")
    print("Model saved to 'models/' directory")
    print("="*80)


if __name__ == '__main__':
    main()

