requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep learning and transformers
transformers>=4.20.0
tokenizers>=0.12.0

# Graph neural networks
torch-geometric>=2.0.0
torch-scatter>=2.0.9
torch-sparse>=0.6.12

# Federated learning
syft>=0.7.0
opacus>=1.3.0

# Data processing
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
pyyaml>=5.4.0
tqdm>=4.62.0
tensorboard>=2.8.0
wandb>=0.12.0

# Privacy and security
cryptography>=36.0.0
pycryptodome>=3.15.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

